backend/requirements.txt
aiofiles==25.1.0
aioimaplib==2.0.1
alembic==1.18.0
asyncpg==0.31.0
black==25.12.0
fastapi==0.128.0
httpx==0.28.1
psycopg2-binary==2.9.11
pydantic[email]==2.12.5
pydantic-settings==2.12.0
PyJWT==2.10.1
PySocks==1.7.1
redis==7.1.0
SQLAlchemy==2.0.45
uvicorn==0.40.0
